+++ src/main/java/ru/sgu/tariff/Service/TariffService.java
@@ -5,7 +5,9 @@ import ru.sgu.tariff.Repository.TariffsRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.List;
+import java.util.function.Supplier;
 
 @Service
 public class TariffService {
@@ -17,23 +19,37 @@ public class TariffService {
         this.tariffsRepository = tariffsRepository;
     }
 
-    public Tariffs create(Tariffs tariff){
+    public Tariffs create(Tariffs tariff) {
+        tariff.setLastModifiedDate(new Date());
         return tariffsRepository.save(tariff);
     }
 
-    public List<Tariffs> findAll(){
+    public List<Tariffs> findAll() {
         return tariffsRepository.findAll();
     }
 
-    public Tariffs find(Long id){
-        return tariffsRepository.findById(id).orElse(null);
+    public Tariffs find(Long id) throws ThereIsNoSuchTariffException {
+        return tariffsRepository.findById(id).orElseThrow(ThereIsNoSuchTariffException::new);
     }
 
-    public Tariffs update(Tariffs tariff){
-        return tariffsRepository.save(tariff);
+    public Tariffs update(Long id, Tariffs tariff) throws ThereIsNoSuchTariffException {
+        Tariffs oldTariff = find(id);
+        if (oldTariff == null) {
+            throw new ThereIsNoSuchTariffException();
+        }
+        if (tariff.getBid() != 0) {
+            oldTariff.setBid(tariff.getBid());
+            oldTariff.setLastModifiedDate(new Date());
+        }
+        return tariffsRepository.save(oldTariff);
     }
 
-    public void delete(Long id){
-        tariffsRepository.deleteById(id);
+    public void delete(Long id) throws ThereIsNoSuchTariffException {
+        int f = tariffsRepository.findById(id).isEmpty() ? 1 : 0;
+        if (f == 0) {
+            tariffsRepository.deleteById(id);
+        } else throw new ThereIsNoSuchTariffException();
     }
+
+    public static class ThereIsNoSuchTariffException extends Exception {}
 }
